+++ R/pip_find_data.R
@@ -1,6 +1,7 @@
+# @description `r lifecycle::badge("deprecated")`
+
 #' Find surveys available in PIP drive
 #'
-#' @description `r lifecycle::badge("deprecated")`
 #'
 #' This function was deprecated because DLW data will loaded directly from a
 #' flat folder structure which allows bypassing the use of `datalibweb` in
@@ -94,9 +95,9 @@ pip_find_data <- function(country         = NULL,
                          verbose      = getOption("pipload.verbose")
                          ) {
 
-  lifecycle::deprecate_soft("0.1.13",
-                            "pip_find_data()",
-                            "pip_find_dlw()")
+  # lifecycle::deprecate_soft("0.1.13",
+  #                           "pip_find_data()",
+  #                           "pip_find_dlw()")
 
 
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~