+++ tests/testthat/test-pip_create_globals.R
@@ -1,0 +1,19 @@
+test_that("wrong names in vintage", {
+
+  expect_error(pip_create_globals(vintage = list(test = "hola", test2 = "chao")))
+
+})
+
+
+# test_that("only accept lists and characters of length 1", {
+#
+# })
+#
+# test_that("inputs within vintage lists are correct", {
+#
+#
+#
+#
+# })
+#
+#